+++ CoVAR-app/src/App.tsx
@@ -1,14 +1,52 @@
 import React from 'react';
+import LoginSignup from './loginSignup';
+import { createTheme } from '@mui/material/styles';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Login from './login';
 
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#52796F',
+    },
+    background: {
+      default: '#1F282E',
+      paper: '#1F282E',
+    },
+    text: {
+      primary: '#CAD2C5',
+      secondary: '#84A98C',
+    },
+  },
+  typography: {
+    fontFamily: 'Urbanist, Arial, sans-serif',
+    button: {
+      textTransform: 'none',
+    },
+  },
+  components: {
+    MuiInputBase: {
+      styleOverrides: {
+        input: {
+          color: '#CAD2C5',
+        },
+      },
+    },
+    MuiFormLabel: {
+      styleOverrides: {
+        root: {
+          color: '#CAD2C5',
+        },
+      },
+    },
+  },
+});
 
 const App: React.FC = () => {
   return (
     <Router>
       <div>
         <Routes>
-          <Route path="/login" element={<Login />} />
+          <Route path="/loginSignup" element={< LoginSignup />} />
           <Route path="/" element={<h2>Home Page</h2>} />
         </Routes>
       </div>
@@ -17,3 +55,4 @@ const App: React.FC = () => {
 };
 
 export default App;
+export { theme };
