+++ CoVAR-app/src/loginSignup.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Login from "./loginForm";
+import Signup from "./signupForm";
+
+const loginSignup = () => {
+    return (
+        < Login />
+        // < Signup />
+    );
+};
+
+export default loginSignup;
