+++ package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "@mui/icons-material": "^5.15.19",
+    "react-router-dom": "^6.23.1"
+  }
+}
